single listen script
